Distinguish between "tab size" and "indent size"
This patch splits the current "tab size" setting into two separate
settings:

* Tab size   : determines the width of the ASCII `TAB` character.
* Indent size: determines the tab stop width.

The first option defaults to 8, the latter to 4.  Additionally, "Replace
by spaces" is renamed to "Use only spaces", to better reflect what it
actually does.

Pressing the TAB key simply inserts enough space characters to move the
cursor to the next tab stop.  But if "Use only spaces" is unchecked, and
the cursor column is divisible by the tab character size, those spaces
are replaced by tabs.  This doesn't even require any extra code, it is
a standard Scintilla feature.

For example, with tab size 8 and indent size 4, pressing the TAB key
on an empty line inserts four spaces.  Pressing it again replaces those
four spaces with a single `TAB` character.

When tab size and indent size are equal, the behavior is the same as
before.

Fixes: #10882
Fixes: #3382